Ensure load balancer does not attempt to add existing master node to cluster.
As documented here:
https://github.com/datacratic/StarCluster/issues/44

The load balancer can mistakenly attempt to add an existing master node
to a cluster, which can introduce problems for run-once configurations
that occur on_add_node(). It's possible that in an intermittent lapse of
connectivity the status is marked as not running, which would include
the master node in the list of instances to add. Prevent this from
happening by checking the name of the instance before adding to the
cluster.